backend/requirements.txt
Django==4.2.3
django-cors-headers==4.2.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
Pillow==9.3.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
python-dotenv==1.0.1
mysqlclient==2.2.4
PyMySQL==1.1.1
psycopg2-binary==2.9.9